simple_forum_v1.0+/requirements.txt
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Admin==1.6.0
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.1
SQLAlchemy==1.4.37
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0